sync leetcode codes 2022-12

+++ leetcode/1-1000/391-perfect-rectangle.cpp
@@ -8,13 +8,13 @@ public:
             if(!Hash.insert(msk).second)
                 Hash.erase(msk);
         };
-        int sum = 0;
+        ULL sum = 0;
         int xL = INT_MAX, xR = INT_MIN;
         int yL = xL, yR = xR;
         for(auto &it : rectangles) {
             int x0, y0, x1, y1;
             tie(x0, y0, x1, y1) = tie(it[0], it[1], it[2], it[3]);
-            sum += (x1 - x0) * (y1 - y0);
+            sum += (ULL)(x1 - x0) * (ULL)(y1 - y0);
             xL = min(xL, x0);
             yL = min(yL, y0);
             xR = max(xR, x1);
@@ -28,6 +28,6 @@ public:
         upd(xL, yR);
         upd(xR, yL);
         upd(xR, yR);
-        return sum == (xR - xL) * (yR - yL) && Hash.empty();
+        return sum == (ULL)(xR - xL) * (ULL)(yR - yL) && Hash.empty();
     }
 };
